+++ src/JavaScriptRuntime/Int32Array.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Buffers.Binary;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace JavaScriptRuntime
 {
@@ -8,15 +9,18 @@ namespace JavaScriptRuntime
     public class Int32Array : TypedArrayBase
     {
         private const int ElementSize = 4;
+        private readonly bool _hasFixedContiguousBacking;
 
         public Int32Array()
         {
             InitializeEmpty();
+            _hasFixedContiguousBacking = HasFixedContiguousBacking();
         }
 
         public Int32Array(object? arg)
         {
             InitializeFromArgument(arg);
+            _hasFixedContiguousBacking = HasFixedContiguousBacking();
         }
 
         public Int32Array(object? arg, object? byteOffset)
@@ -24,10 +28,12 @@ namespace JavaScriptRuntime
             if (arg is ArrayBuffer arrayBuffer)
             {
                 InitializeFromBuffer(arrayBuffer, byteOffset, null);
+                _hasFixedContiguousBacking = HasFixedContiguousBacking();
                 return;
             }
 
             InitializeFromArgument(arg);
+            _hasFixedContiguousBacking = HasFixedContiguousBacking();
         }
 
         public Int32Array(object? arg, object? byteOffset, object? length)
@@ -35,15 +41,18 @@ namespace JavaScriptRuntime
             if (arg is ArrayBuffer arrayBuffer)
             {
                 InitializeFromBuffer(arrayBuffer, byteOffset, length);
+                _hasFixedContiguousBacking = HasFixedContiguousBacking();
                 return;
             }
 
             InitializeFromArgument(arg);
+            _hasFixedContiguousBacking = HasFixedContiguousBacking();
         }
 
         private Int32Array(ArrayBuffer buffer, int byteOffset, int length)
         {
             InitializeFromExisting(buffer, byteOffset, length);
+            _hasFixedContiguousBacking = HasFixedContiguousBacking();
         }
 
         public static Int32Array from(object? source)
@@ -75,8 +84,35 @@ namespace JavaScriptRuntime
 
         public new double this[double index]
         {
-            get => base[index];
-            set => base[index] = value;
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get
+            {
+                var candidate = (int)index;
+                if (candidate == index
+                    && (uint)candidate < (uint)FixedLengthElements
+                    && HasContiguousBacking)
+                {
+                    return MemoryMarshal.Cast<byte, int>(BufferObject.RawBytes)[
+                        (ByteOffsetBytes / ElementSize) + candidate];
+                }
+
+                return base[index];
+            }
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            set
+            {
+                var candidate = (int)index;
+                if (candidate == index
+                    && (uint)candidate < (uint)FixedLengthElements
+                    && HasContiguousBacking)
+                {
+                    var elements = MemoryMarshal.Cast<byte, int>(BufferObject.RawBytes.AsSpan());
+                    elements[(ByteOffsetBytes / ElementSize) + candidate] = ToInt32(value);
+                    return;
+                }
+
+                base[index] = value;
+            }
         }
 
         public Int32Array slice()
@@ -99,6 +135,11 @@ namespace JavaScriptRuntime
 
         protected override double ReadElementValue(int index)
         {
+            if (TryGetContiguousElements(out var elements))
+            {
+                return elements[index];
+            }
+
             var offset = ByteOffsetBytes + (index * ElementSize);
             return ReadInt32(BufferObject.RawBytes, offset);
         }
@@ -106,9 +147,37 @@ namespace JavaScriptRuntime
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         protected override void WriteElementValue(int index, double value)
         {
+            if (TryGetContiguousElements(out var elements))
+            {
+                elements[index] = ToInt32(value);
+                return;
+            }
+
             var offset = ByteOffsetBytes + (index * ElementSize);
             WriteInt32(BufferObject.RawBytes, offset, ToInt32(value));
         }
+
+        // The returned span is only valid until the next operation that can change the buffer.
+        // Callers must not retain it across JavaScript execution or a buffer transfer.
+        internal bool TryGetContiguousElements(out Span<int> elements)
+        {
+            if (!HasContiguousBacking)
+            {
+                elements = default;
+                return false;
+            }
+
+            elements = MemoryMarshal.Cast<byte, int>(
+                BufferObject.RawBytes.AsSpan(ByteOffsetBytes, LengthElements * ElementSize));
+            return true;
+        }
+
+        private bool HasContiguousBacking => _hasFixedContiguousBacking && !BufferObject.IsDetached;
+
+        private bool HasFixedContiguousBacking()
+            => !BufferObject.IsResizable
+                && BufferObject is not SharedArrayBuffer { growable: true }
+                && (ByteOffsetBytes % ElementSize) == 0;
 
         protected override TypedArrayBase CreateSameType(ArrayBuffer buffer, int byteOffset, int length)
             => new Int32Array(buffer, byteOffset, length);